armando mainlayout
