+++ src/Pages/Home/home.scss
@@ -1,0 +1,34 @@
+.Home {
+    text-align: center;
+}
+  
+.App-logo {
+    height: 40vmin;
+    pointer-events: none;
+}
+
+.TrendingTitle{
+    display: flex;
+    align-items: center;  
+    flex-direction: row;
+    margin-top: 15px;
+    margin-left: 15px;
+    font-weight: 600;
+  }
+  
+  .titleTrending{
+    margin-left: 10px;
+  }
+  
+  .TrendingPostsContianer{
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: center;
+    border-bottom: 1px solid rgb(242,242,242);
+  }
+  
+  .Content-Body{
+    display: flex;
+    flex-direction: row;
+  }
